Adjust for SLE

+++ 50-tools/prepare_split_tw_leap.sh
@@ -6,10 +6,6 @@ set -o pipefail
 
 export LANG=C.utf8
 
-## get package lists
-rm -rf 50-lists
-mkdir 50-lists
-
 # die: Echo arguments to stderr and exit with 1
 die() { echo "$@" 1>&2 ; exit 1; }
 log()
@@ -35,7 +31,9 @@ move_if_changed()
 	fi
 	if [ "$updated" != 0 ]; then
 	  mv "$src" "$dst"
-	  msgfmt --statistics -o /dev/null "$dst"
+	  if [ "$VERBOSE" = 1 ]; then
+	    msgfmt --statistics -o /dev/null "$dst"
+	  fi
 	fi
 }
 
@@ -45,6 +43,9 @@ fi
 urlfile="$1"
 shift
 
+# get package lists
+rm -rf 50-lists
+mkdir 50-lists
 while read distro url; do
   case $distro in
     \#*) continue ;;
@@ -65,20 +66,26 @@ msgcat *._pot | grep -v "#-#-#-#" > _packages.pot
 #cat _packages.pot | msggrep -X -i -e "^tumbleweed/patterns\|^leap/patterns" -o patterns.pot
 #msgcat --unique -o __packages.pot patterns.pot _packages.pot && mv __packages.pot _packages.pot
 
-echo "Splitting started: $(date)"
+log "Splitting started: $(date)"
 for i in patterns {a..z}; do
   I=${i^^}
   log -n "$i "
-  cat _packages.pot | msggrep -X -e "^tumbleweed/$i\|^leap/$i\|^tumbleweed/$I\|^leap/$I" -o $i.pot
-  msgcat --unique -o __packages.pot $i.pot _packages.pot && mv __packages.pot _packages.pot
+  cat _packages.pot | msggrep -X -e "^[^\/]\+/\($i\|$I\)" -o $i.pot
+  if [ -e "$i.pot" ]; then
+    msgcat --unique --force-po -o __packages.pot $i.pot _packages.pot && mv __packages.pot _packages.pot
+  fi
 done
+# move all the remaining packges in a
 tail -n +8 a.pot >> _packages.pot && mv _packages.pot a.pot
 
 for ii in aspell ghc gnome golang google gstreamer gtk kde leechcraft libreoffice libqt lib mate myspell perl php python rubygem tesseract texlive-specs texlive wx xfce4 yast2; do
   firstChar=${ii:0:1}
+  [ -e "$fistChar.pot" ] || continue
   log -n "$ii "
-  msggrep -X -e "^tumbleweed/$ii\|^leap/$ii" "$firstChar.pot" -o $ii.pot --no-wrap
-  msgcat --unique -o _$firstChar.pot $ii.pot $firstChar.pot && mv _$firstChar.pot $firstChar.pot
+  msggrep -X -e "^[^/]\+/$ii" "$firstChar.pot" -o $ii.pot --no-wrap
+  if [ -e "$ii.pot" ]; then
+    msgcat --unique -o _$firstChar.pot $ii.pot $firstChar.pot && mv _$firstChar.pot $firstChar.pot
+  fi
 done
 log ""
 log "Splitting finished: $(date)"
